价格调整

+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/pojo/ApplyProductSkuCheckout.java
@@ -20,19 +20,19 @@ public class ApplyProductSkuCheckout extends CommonBean {
     private String skuName;
 
     @ApiModelProperty("销项税率")
-    private Long outputTaxRate;
+    private BigDecimal outputTaxRate;
 
     @ApiModelProperty("进项税率")
-    private Long inputTaxRate;
+    private BigDecimal inputTaxRate;
 
     @ApiModelProperty("积分系数")
-    private Long integralCoefficient;
+    private BigDecimal integralCoefficient;
 
     @ApiModelProperty("申请编码")
     private String applyCode;
 
     @ApiModelProperty("返点")
-    private Long rebate;
+    private BigDecimal rebate;
 
     @ApiModelProperty("物流费奖励比例")
     private BigDecimal logisticsFeeAwardRatio;
